Example 3 2

+++ src/main/java/com/example/demo/ex3/Example3.java
@@ -1,13 +1,13 @@
 package com.example.demo.ex3;
 
 public class Example3 {
-    int a = 10;
+    long a = 10;
 
-    int b = 20;
+    long b = 20;
 
     int c = 30;
 
-    int d = 40;
+    long d = 40;
 
-    int e = 50;
+    long e = 50;
 }
